+++ src/cases/case-feature.routes.ts
@@ -15,7 +15,6 @@ import { CaseFilterComponent } from './containers/case-filter/case-filter.compon
 import { CaseHomeComponent } from './containers/case-home/case-home.component';
 import { CaseListComponent } from './containers/case-list/case-list.component';
 import { CaseSearchComponent } from './containers/case-search/case-search.component';
-import { RoleAccessComponent } from './containers/role-and-access/role-access.component';
 import { ActivityResolver } from './resolvers/activity.resolver';
 import { CreateCaseEventTriggerResolver } from './resolvers/create-case-event-trigger.resolver';
 
@@ -108,10 +107,6 @@ export const ROUTES: Routes = [
           data: {
             title: 'Case Details'
           }
-        },
-        {
-          path: 'role-access',
-          component: RoleAccessComponent
         }
       ]
     },
