+++ src/cases/services/index.ts
@@ -1,7 +1,5 @@
-import { RoleExclusionsService } from './role-exclusions.service';
 import { SearchFilterService } from './search-filter.service';
 
-export const services: any[] = [SearchFilterService, RoleExclusionsService];
+export const services: any[] = [SearchFilterService];
 
 export * from './search-filter.service';
-export * from './role-exclusions.service';
