+++ src/role-access/components/exclusion/index.ts
@@ -1,7 +1,9 @@
 import { DescribeExclusionComponent } from './describe-exclusion/describe-exclusion.component';
 import { AllocateRoleComponent } from './allocate-role/allocate-role.component';
+import { RoleExclusionsComponent } from './role-exclusions/role-exclusions.component';
 
-export const components: any[] = [ AllocateRoleComponent, DescribeExclusionComponent];
+export const components: any[] = [ AllocateRoleComponent, DescribeExclusionComponent, RoleExclusionsComponent];
 
 export * from './describe-exclusion/describe-exclusion.component';
 export * from './allocate-role/allocate-role.component';
+export * from './role-exclusions/role-exclusions.component';
