+++ src/role-access/containers/index.ts
@@ -5,6 +5,7 @@ import { DescribeExclusionComponent } from './exclusion/describe-exclusion/descr
 import { ExclusionHomeComponent } from './exclusion/exclusion-home/exclusion-home.component';
 import { ExclusionNavigationComponent } from './exclusion/exclusion-navigation/exclusion-navigation.component';
 import { FindPersonComponent } from './exclusion/find-person/find-person.component';
+import { RoleAccessComponent } from './role-and-access/role-access.component';
 
 export const containers: any[] = [
   ExclusionHomeComponent,
@@ -13,7 +14,8 @@ export const containers: any[] = [
   ChoosePersonRoleComponent,
   FindPersonComponent,
   DescribeExclusionComponent,
-  CheckAnswersComponent
+  CheckAnswersComponent,
+  RoleAccessComponent
 ];
 
 export * from './exclusion/check-answers/check-answers.component';
@@ -23,3 +25,4 @@ export * from './exclusion/describe-exclusion/describe-exclusion.component';
 export * from './exclusion/exclusion-home/exclusion-home.component';
 export * from './exclusion/exclusion-navigation/exclusion-navigation.component';
 export * from './exclusion/find-person/find-person.component';
+export * from './role-and-access/role-access.component';
