+++ src/role-access/containers/role-and-access/role-access.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { RoleExclusion } from '../../../cases/models/role-exclusions/role-exclusion.model';
-import { RoleExclusionsService } from '../../../cases/services';
+import { RoleExclusion } from '../../models/role-exclusion.model';
+import { RoleExclusionsService } from '../../service/role-exclusions.service';
 
 @Component({
     selector: 'exui-role-access',
