+++ src/role-access/models/index.ts
@@ -4,3 +4,4 @@ export * from './exclusion-state-data.interface';
 export * from './http-error.interface';
 export * from './exclusion.enum';
 export * from './person-model';
+export * from './role-exclusion.model';
