+++ src/role-access/service/index.ts
@@ -1,0 +1,5 @@
+import { RoleExclusionsService } from './role-exclusions.service';
+
+export const services: any[] = [RoleExclusionsService];
+
+export * from './role-exclusions.service';
